a stub's value always wins over the template's
In the 99% case of providing sane defaults in a template, you also want to
allow the stubs to override it. This has led to a lot of repeated "merge || 1"
type expressions, which is very noisy and cumbersome.

For simple values (integers, strings, nil), a stub's value will now always
win. For lists and maps, you must be explicit, as the semantics are more
complicated (do you want to insert or replace into a list, for example).