We care about order when using the <<: (( merge )) operation.
Previously we got this for free, but when go 1.3 made map iteration random it
broke for us.  So just sort the keys before interating.

Another option is to specifically pick out the << key and run it first.

Since the behavior is random, you must run the changes made in this commit to
the "map splicing" test without the fix several times, or just add a loop
around the test expectation, to reproduce the issue.