support multiplication and grammer rework
multiplication + division has been added. The concatenation expression
is extended to concatenate any int/string sequence.

Additionally the grammer has been tuned, to correctly handle white spaces
in lists, to handle chained subtraction (no right recursion ( 6-3-2 = 1 not 5)) and to fix
the evaluation order for concatenation. Now it is possible to use
expressions as part of concatenations.