general arithmetic expression fix
The current implementation cannot handle incremental expression evaluation correctly.

a: 1
b: 2
c: (( b ))
plus: (( a + b || "failed" ))

always resolves to "failed", because of the required delayed evaluation of c.
In general all expressions must be checked to handle this case correctly.
Here comes the master fix for the arithmetic expressions.