fix for defaulted static_ips based on expressions
if static_ips is used together with the || operator
and an implicitly used value from the network or job
definition is using a dynaml expression, the
resulting value is always the default.

The calculation of the static ips must be delayed until
the indirectly used elements are completely resolved,
instead of delivering an error. Therefore the
reference evaluation now uses reference expressions,
instead of a direct lookup in the yaml document.
This provides the opportunity to handle the unresolved case.